+++ libs/__init__.py
@@ -1,14 +1,15 @@
 # pylint:disable=wrong-import-position
 """库文件"""
 __all__ = [
-    "START", "OUTPUT", "get_user_info", "music_list_dispatcher", 
-    "DrawBest", "Config", "load_config"
+    "START", "OUTPUT", "music_list_dispatcher", 
+    "DrawBest", "Config", "load_config", "get_user_data",
 ]
 
 import time
 START = time.time()
 
 from .consts import OUTPUT
-from .api import get_user_info, music_list_dispatcher
+from .api import music_list_dispatcher
 from .image import DrawBest
 from .config import Config, load_config
+from .parser import get_user_data